chore: Update index.html and vite.config.ts
This commit updates the `index.html` file by changing the page title to "Controle Produtos". Additionally, the `vite.config.ts` file is modified to add a proxy configuration for API requests. The proxy redirects requests starting with "/api" to the local server at "http://localhost:8000". This change improves the functionality and integration of the frontend with the backend API.